fixbug

+++ src/views/principal/certification/list.vue
@@ -506,7 +506,20 @@ export default {
         },
         // 删除多个
         deleteHandler() {
-            deleteCertificationsAction(this);
+            this.$Modal.confirm({
+                title: '提示',
+                content: '确定要删除吗',
+                okText: '确定',
+                cancelText: '取消',
+                onOk: () => {
+                    deleteCertificationsAction(this);
+                },
+                onCancel: () => {
+                    
+                    
+                }
+            })
+            
         },
     
         handleEdit(row){
