+++ src/views/reports/login-list/list.vue
@@ -277,7 +277,20 @@ export default {
         },
         // 删除多个
         deleteHandler() {
-            deleteLoginRecordAction(this);
+            this.$Modal.confirm({
+                title: '提示',
+                content: '确定要删除吗',
+                okText: '确定',
+                cancelText: '取消',
+                onOk: () => {
+                    deleteLoginRecordAction(this);
+                },
+                onCancel: () => {
+                    
+                    
+                }
+            })
+            
         },
         
     },
