+++ src/views/reports/unlock-list/list.vue
@@ -287,7 +287,20 @@ export default {
         },
         // 删除多个
         deleteHandler() {
-            deleteUnlockAction(this);
+            this.$Modal.confirm({
+                title: '提示',
+                content: '确定要删除吗',
+                okText: '确定',
+                cancelText: '取消',
+                onOk: () => {
+                    deleteUnlockAction(this);
+                },
+                onCancel: () => {
+                    
+                    
+                }
+            })
+            
         },
         
     },
